+++ hooks/pre_gen_project.py
@@ -1,32 +1,57 @@
+import string
+import sys
+import warnings
+
+
 def check_repo_name_structure(repo_name: str) -> None:
-    """Check if the repository name follows the correct structure.
+    """Check if the repo name follows the correct structure.
 
     Args:
-        repo_name: The name of the repository to check.
+        repo_name: The name of the repo to check.
 
     Raises:
-        ValueError: If the repository name does not follow the correct structure.
+        ValueError: If the repo name does not follow the correct structure.
     """
     if len(repo_name) > 88:
-        print("checking repo name length")
-        raise ValueError("Repository name must not exceed 88 characters.")
+        raise ValueError(
+            "Repo name must not exceed 88 characters."
+        )
 
+    allowed_characters = string.ascii_lowercase + string.digits + "_"
+    if any(char not in allowed_characters for char in repo_name):
+        raise ValueError(
+            "Repo name contains invalid characters. Only lowercase letters, "
+            "digits, and underscores are allowed."
+        )
+
+    if "_" in repo_name:
+        if repo_name.startswith("_"):
+            raise ValueError(
+                "Repo name must not start with an underscore."
+            )
+        elif repo_name.endswith("_"):
+            raise ValueError(
+                "Repo name must not end with an underscore."
+            )
+        else:
+            warnings.warn(
+                "Underscores are discouraged in Python package " +
+                "names unless they improve readability.",
+                UserWarning
+            )
 
 if __name__ == "__main__":
 
-    {
-        {
-            cookiecutter.update(  # noqa: F821
-                {
-                    "project_slug": cookiecutter.project_name.lower()  # noqa: F821
-                    | replace(" ", "_")  # noqa: F821
-                    | replace("-", "_")  # noqa: F821
-                    | replace(".", "_")  # noqa: F821
-                    | trim()  # noqa: F821
-                }
-            )
-        }
-    }
+    {{ cookiecutter.update({
+        "project_slug": cookiecutter.project_name.lower()
+        |replace(' ', '_')
+        |replace('-', '_')
+        |replace('.', '_')
+        |trim()
+        })
+    }}
 
     # Check the format of the contact email address supplied is a valid one
     check_repo_name_structure("{{ cookiecutter.project_slug }}")
+
+    sys.exit(0)